feat: adds a classes for reporting results to the console
Abstracts reporting results to the console to different class
and out of the entrypoint base class to allow it to extend more easily
and be used in different places in the code base.
