+++ trestlebot/entrypoints/entrypoint_base.py
@@ -18,10 +18,11 @@ import sys
 from typing import List, Optional
 
 from trestlebot import const
-from trestlebot.bot import BotResults, TrestleBot
+from trestlebot.bot import TrestleBot
 from trestlebot.github import GitHub, is_github_actions
 from trestlebot.gitlab import GitLab, get_gitlab_root_url, is_gitlab_ci
 from trestlebot.provider import GitProvider
+from trestlebot.reporter import BotResults, ResultsReporter
 from trestlebot.tasks.base_task import TaskBase
 
 
@@ -177,17 +178,9 @@ class EntrypointBase:
             dry_run=args.dry_run,
         )
 
-        # Print results to stdout
-        if results.changes:
-            print("Changes:")  # noqa: T201
-            for change in results.changes:
-                print(change)  # noqa: T201
-
-        if results.commit_sha:
-            print(f"Commit Hash: {results.commit_sha}")  # noqa: T201
-
-        if results.pr_number:
-            print(f"Pull Request Number: {results.pr_number}")  # noqa: T201
+        # Report the results to the console
+        reporter = ResultsReporter()
+        print(reporter.report_results(results))  # noqa: T201
 
 
 def comma_sep_to_list(string: str) -> List[str]:
